moves mood value outside of the button component
