+++ src/common-components/CommonComponents.css
@@ -1,100 +1,111 @@
-
+* {
+  --mood-value-diameter: 45px;
+  --mood-border-thickness: 3px;
+}
 /*BUTTONS*/
 .fab-large {
-	color: var(--accent-2-base);
-	background-color: var(--background-highlight);
-	font-size: 3em;
-	padding: 0.7em 1em;
-	border-radius: 0.5em;
-	box-shadow: 0 4px 4px 2px rgba(0, 0, 0, 0.25);
-	border: none;
-  }
-  .fab-large:hover {
-	background-color: var(--accent-2-base);
-	color: var(--background-highlight);
-	box-shadow: 0 4px 4px 2px rgba(0, 0, 0, 0.25);
+  color: var(--accent-2-base);
+  background-color: var(--background-highlight);
+  font-size: 3em;
+  padding: 0.7em 1em;
+  border-radius: 0.5em;
+  box-shadow: 0 4px 4px 2px rgba(0, 0, 0, 0.25);
+  border: none;
+}
+.fab-large:hover {
+  background-color: var(--accent-2-base);
+  color: var(--background-highlight);
+  box-shadow: 0 4px 4px 2px rgba(0, 0, 0, 0.25);
+}
+.filled-button {
+  background-color: var(--accent-base);
+  color: var(--foreground-contrast);
+  font-size: 1.5em;
+  padding: 6px auto;
+  border-radius: 0.5em;
+  border: none;
+}
+.filled-button:hover {
+  background-color: var(--accent-highlight);
+}
+.filled-button:active {
+  background-color: var(--accent-shadow);
+}
+.filled-button:disabled {
+  background-color: var(--gray-highlight);
+  color: var(--gray-base);
+}
+.icon-button {
+  border: 3px solid transparent;
+  background-color: transparent;
+  padding: 2px 4px;
+}
+.icon-button img {
+  height: 3em;
+}
+.icon-button:hover {
+  border-color: var(--transparent-black);
+}
+@media screen and (width < 599px) {
+  .filled-button-container {
+    background-color: var(--background-base);
+    width: 100dvw;
+    position: fixed;
+    right: 0px;
+    bottom: 0px;
+    text-align: center;
   }
   .filled-button {
-	background-color: var(--accent-base);
-	color: var(--foreground-contrast);
-	font-size: 1.5em;
-	padding: 6px auto;
-	border-radius: 0.5em;
-	border: none;
+    margin: 16px 0;
+    width: 70dvw;
   }
-  .filled-button:hover {
-	background-color: var(--accent-highlight);
-  }
-  .filled-button:active {
-	background-color: var(--accent-shadow);
-  }
-  .filled-button:disabled {
-	background-color: var(--gray-highlight);
-	color: var(--gray-base);
-  }
-  .icon-button{
-	border: 3px solid transparent;
-	background-color: transparent;
-	padding:2px 4px;
-  }
-  .icon-button img{
-	height: 3em;
-  }
-  .icon-button:hover{
-	border-color: var(--transparent-black);
-  }
-  @media screen and (width < 599px) {
-	.filled-button-container{
-	  background-color: var(--background-base);
-	  width:100dvw;
-	  position: fixed;
-	  right: 0px;
-	  bottom: 0px;
-	  text-align: center;
-	}
-	.filled-button
-	{
-		margin: 16px 0;
-	  width: 70dvw;
-	}
-  }
-/*MOOD VALUE BUTTONS*/
-.mood-value.selected{
-	border: 3px solid var(--transparent-black);
 }
-.mood-value{
-	flex-shrink: 0;
-	border:none;
-	padding: 0;
-	height: 45px;
-	width: 45px;
-	border-radius: 50%;
-	
-	background-repeat: no-repeat;
-	background-position: center;
+/*MOOD VALUE BUTTONS*/
+.mood-value-button {
+  flex-shrink: 0;
+  border: none;
+  padding: 0;
+  height: var(--mood-value-diameter);
+  width: var(--mood-value-diameter);
+  border-radius: 50%;
+}
+.selected-value .mood-value {
+  border: var(--mood-border-thickness) solid var(--transparent-black);
+  height: calc(var(--mood-value-diameter) - (var(--mood-border-thickness) * 2));
+  width: calc(var(--mood-value-diameter) - (var(--mood-border-thickness) * 2));
+}
+.mood-value {
+  flex-shrink: 0;
+  padding: 0;
+  height: var(--mood-value-diameter);
+  width: var(--mood-value-diameter);
+  border-radius: 50%;
+
+  background-repeat: no-repeat;
+  background-position: center;
 }
 /*TAGS*/
 .tag {
-	display: flex;
-	align-items: center;
-	column-gap: 4px;
-	padding: 2px 10px;
-	border-radius: 0.5em;
-	box-shadow: 2px 3px 3px 2px rgba(0, 0, 0, 0.25);
-	background-color: white;
-  }
-  .tag > div {
-	width: 0.5em;
-	height: 0.5em;
-	border-radius: 0.5em;
-	margin-right: 4px;
-  }
-  .tag-container{
-	display: flex;
-	align-items: center;
-	justify-content: flex-end;
-	flex-wrap: wrap;
-	flex-grow:1;
-	column-gap: 8px;
-	row-gap: 8px;
-  }
+  display: flex;
+  align-items: center;
+  column-gap: 4px;
+  padding: 2px 10px;
+  border-radius: 0.5em;
+  box-shadow: 2px 3px 3px 2px rgba(0, 0, 0, 0.25);
+  background-color: white;
+}
+.tag > div {
+  width: 0.5em;
+  height: 0.5em;
+  border-radius: 0.5em;
+  margin-right: 4px;
+}
+.tag-container {
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+  flex-wrap: wrap;
+  flex-grow: 1;
+  column-gap: 8px;
+  row-gap: 8px;
+}
