+++ src/data/apiMock.ts
@@ -31,7 +31,6 @@ export async function deleteMood(
 
 export async function returnLatestMoods() {
   //returns moods ordered by date-time
-  console.log("fetching");
   await delay(1500);
   return returnLatestMoodsMock();
 }
